server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
cryptography==41.0.7
pyyaml==6.0.1
jinja2==3.1.2
aiofiles==23.2.1
python-multipart==0.0.6
requests==2.31.0
